+++ changelog.rst
@@ -1,6 +1,7 @@
 Change Log
-===========
+----------
 
 0.2.1 released 2011-05-19
------------------------------
- - change SQLAlchemy requirement so the latest package < 0.7 is installed
+=========================
+
+* change SQLAlchemy requirement so the latest package < 0.7 is installed